refactor(observeOn): move support types under observeOn.ts
